Добавлена пагинация для запросов, в которых она требуется. Убран уровень логирования Trace. Сущности из БД больше не используются в контроллерах: вместо них теперь используются новые DTO с аналогочными полями. Добавлены соответствующие методы в Mapper классах. Из EventRequestStatusUpdateResult убран конструктор, вместо него теперь поля проинициализированы сами. В EventShortDto и EventFullDto поле id отправлено вверх. NewCompilationDto: поле events переименовано в eventIds. NewEventDto: поле category переименовано в categoryId. ParticipationRequestDto: id сдвинуто вверх. В классах model убраны лишние аннотации, отвечающие за генерацию конструкторов. Убраны лишние аннотации Transactional. Убран лишний Group By в запросе в CompEventRepository. В EndpointHit убраны аннотации Column.

+++ main/src/main/java/ru/practicum/controller/AdminEventController.java
@@ -34,8 +34,8 @@ public class AdminEventController {
                 @RequestParam(required = false) List<Long> categories,
                 @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime rangeStart,
                 @RequestParam(required = false) @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime rangeEnd,
-                @PositiveOrZero @RequestParam(name = "from", defaultValue = "0") Long from,
-                @Positive @RequestParam(name = "size", defaultValue = "10") Long size) {
+                @PositiveOrZero @RequestParam(name = "from", defaultValue = "0") Integer from,
+                @Positive @RequestParam(name = "size", defaultValue = "10") Integer size) {
         List<EventFullDto> response = adminEventService.getFilteredEvents(
                 users,
                 states,
